Create Class_Functions
This is a rough estimate of how the structure of our code could look like. 

Each class includes what functions it could need.

While we code there will probably be a lot of changes to it, and if someone sees that something is completely off, add comments and then we can go over it on our weekly calls. 

:)