Update graphics.py
Here I implemented the ai_make_move function (it works but needs to be adjusted). At the moment when only minimax is implemented.

Also when the selectStarter() is chosen as Computer the ai isn't making a move, not sure what needs to be done there to make it do smth but when human starts then everything works smoothly.

The points need to be implemented next and the changing the game screen when game is over.

Also implemented the edges visited.

In the terminal when game is going it shows a lot of information that can be used to debug, and gives a better overview of the game as its going.